Desenvolvimento do cadastro do tads
- Criação dos combo-boxes para realização da escolha da cidade e estado no cadastro de usuario

+++ src/main/java/com/dac/tads/dao/CidadeDAO.java
@@ -34,14 +34,14 @@ public class CidadeDAO {
     }
 
     // Retornará um único cidade
-    public Cidade selectCidade(int id) {
+    public Cidade selectCidade(Long id) {
         Cidade cidade = null;
         try {
             Session session = HibernateUtil.getSessionFactory().openSession();
             session.beginTransaction();
             Query query = session.createQuery(
                     "from Cidade where id = :id");
-            query.setInteger("id", id);
+            query.setLong("id", id);
             cidade = (Cidade) query.uniqueResult();
             session.close();
         } catch (HibernateException e) {
